Add id token support (#51)
* Add configuration for authentication provider tokens

* Add support for RSA encrypted tokens

This adds support for RSA encrypted tokens in order to enable compatibility with JWTs signed by authentication providers.

* Accept warrants with no subject for authz checks

This updates the authorize endpoint to accept warrants with no subject when a request includes a session token. Each warrant will be updated to use the user specified by the token.

* Add flag to enable session auth for specific routes

* Fetch user id from claims based on identifier from configuration

Instead of fetching the user id from the standard `sub` field from the JWT claims object, this allows the user id field to be specified if custom claims are used. By default, the user identifier will be `sub`

* Add tenantId support to authInfo

* Add support for tenantId in check service

If a tenantId is populated in authInfo, the Check service will automatically append the tenantId as part of the context for each warrant.

* Update mysql and postgres README to include auth provider config

* Rename UserIdentifier and TenantIdentifier to be more explicit

* Rename AuthProvider to Authentication in config

* Update error responses and add more logging for errors

* Rename import alias from authz to warrant

* Remove IsImplicit from SessionWarrantSpec

* Update environment variable names in READMEs