Update user queries to correctly handle queries where beforeId and/or beforeValue is passed
When a `beforeId` or `beforeValue` is passed along with a limit, the results passed should be the immediate N objects where the `sortBy` value is < `beforeId` or `beforeValue`.

For example, if we have a query with the following params: `?sortBy=id&sortOrder=ASC` we may get the following result set: `[1, 2, 3, 4, 5]`. If we then updated the query to be: `?sortBy=id&sortOrder=ASC&beforeId=4&limit=2`, we should expect the result set to be `[2, 3]` rather than `[1, 2]`.

This commit updates the beforeId/beforeValue cases to invert the sort order when making the original query, then reversing the result set.